# openapi.yaml
openapi: 3.0.3
info:
  title: Financial Records API
  description: |
    REST API for managing financial records with role-based access control,
    JWT authentication, and analytics dashboards.

    ## Authentication
    All endpoints except `/api/auth/register` and `/api/auth/login` require a
    Bearer JWT token in the `Authorization` header.

    ## Response Envelope
    All successful responses are wrapped in a standard envelope:
    ```json
    { "success": true, "data": <payload> }
    ```
    All error responses use:
    ```json
    { "success": false, "error": { "code": "<ERROR_CODE>", "message": "<description>" } }
    ```

    ## Error Codes
    | HTTP Status | Error Code        | Description                          |
    |-------------|-------------------|--------------------------------------|
    | 400         | BAD_REQUEST       | Validation error / bad input         |
    | 400         | FOREIGN_KEY_ERROR | Related record not found (P2003)     |
    | 401         | UNAUTHORIZED      | Missing or invalid JWT token         |
    | 403         | FORBIDDEN         | Insufficient role permissions        |
    | 404         | NOT_FOUND         | Resource not found                   |
    | 404         | RECORD_NOT_FOUND  | Prisma record not found (P2025)      |
    | 409         | CONFLICT          | Duplicate resource (e.g. email)      |
    | 409         | DUPLICATE_ENTRY   | Prisma unique constraint (P2002)     |
...

    ## Role Permissions
...
  version: "1.0.0"
...
servers:
...
security: []
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ───────────────────────────────────────────────────── Reusable schemas
  schemas:
...
    RecordType:
      type: string
      enum: [INCOME, EXPENSE]

    # ── Standard error envelope ──
    ErrorEnvelope:
      type: object
      description: All error responses use this standardized envelope.
      required: [success, error]
      properties:
        success:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
              example: BAD_REQUEST
            message:
              type: string
              description: Human-readable error message
              example: Validation failed

    # ── Auth DTOs ──
...
    CreateRecordRequest:
      type: object
      required: [amount, type, category]
      properties:
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 5000.00
        type:
          $ref: "#/components/schemas/RecordType"
        category:
          type: string
          minLength: 1
          example: salary
        date:
          type: string
          format: date
          description: ISO date string. Defaults to current date.
          example: "2026-04-01"
        notes:
          type: string
          example: April salary payment
    UpdateRecordRequest:
      type: object
      description: All fields from CreateRecordRequest, but optional.
      properties:
        amount:
          type: number
          format: float
          minimum: 0.01
          example: 5500.00
        type:
          $ref: "#/components/schemas/RecordType"
        category:
          type: string
          minLength: 1
          example: salary
        date:
          type: string
          format: date
          example: "2026-04-15"
        notes:
          type: string
          example: Updated salary payment
    RecordCreator:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
    FinancialRecord:
      type: object
      properties:
        id:
          type: integer
          example: 1
        amount:
          type: number
          format: float
          example: 5000.00
        type:
          $ref: "#/components/schemas/RecordType"
        category:
          type: string
          example: salary
        date:
          type: string
          format: date-time
          example: "2026-04-01T00:00:00.000Z"
        notes:
          type: string
          nullable: true
          example: April salary payment
        createdBy:
          type: integer
          example: 1
        createdAt:
          type: string
          format: date-time
          example: "2026-04-01T12:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-04-01T12:00:00.000Z"
        user:
          $ref: "#/components/schemas/RecordCreator"
    PaginatedRecordsData:
      type: object
      description: Inner data for paginated records response
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/FinancialRecord"
        total:
          type: integer
          description: Total records matching the filter
          example: 42
        page:
          type: integer
          description: Current page number
          example: 1
        limit:
          type: integer
          description: Results per page
          example: 10
        totalPages:
          type: integer
          description: Total number of pages
          example: 5

    # ── Dashboard schemas (inner data) ──
...
paths:
...
  /records:
    post:
      tags: [Records]
      summary: Create record (Admin only)
      operationId: createRecord
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRecordRequest"
      responses:
        "201":
          description: Record created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/FinancialRecord"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden - requires ADMIN role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    get:
      tags: [Records]
      summary: List records with filters and pagination (Admin, Analyst)
      operationId: listRecords
      security:
        - BearerAuth: []
      parameters:
        - name: type
          in: query
          required: false
          description: Filter by record type
          schema:
            $ref: "#/components/schemas/RecordType"
        - name: category
          in: query
          required: false
          description: Case-insensitive contains match on category
          schema:
            type: string
            example: salary
        - name: startDate
          in: query
          required: false
          description: Records on or after this date (ISO format)
          schema:
            type: string
            format: date
            example: "2026-01-01"
        - name: endDate
          in: query
          required: false
          description: Records on or before this date (ISO format)
          schema:
            type: string
            format: date
            example: "2026-12-31"
        - name: search
          in: query
          required: false
          description: Search in category or notes
          schema:
            type: string
            example: food
        - name: page
          in: query
          required: false
          description: Page number (default 1)
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          required: false
          description: Results per page (default 10, max 100)
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
      responses:
        "200":
          description: Paginated list of records
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/PaginatedRecordsData"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden - requires ADMIN or ANALYST role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /records/{id}:
    get:
      tags: [Records]
      summary: Get record by id (Admin, Analyst)
      operationId: getRecordById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Record ID
      responses:
        "200":
          description: Record found
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/FinancialRecord"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden - requires ADMIN or ANALYST role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    patch:
      tags: [Records]
      summary: Update record (Admin only)
      operationId: updateRecord
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Record ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRecordRequest"
      responses:
        "200":
          description: Record updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    enum: [true]
                  data:
                    $ref: "#/components/schemas/FinancialRecord"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden - requires ADMIN role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

    delete:
      tags: [Records]
      summary: Delete record (Admin only)
      operationId: deleteRecord
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          description: Record ID
      responses:
        "204":
          description: Record deleted (no content)
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "403":
          description: Forbidden - requires ADMIN role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Record not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ─────────────────────────────────────────────────── Dashboard
...